implement rebase_stack, stop popping so much!

+++ src/ast.rs
@@ -8,14 +8,14 @@ pub type Block = Vec<Arc<Stmt>>;
 
 #[derive(Debug, Clone)]
 pub struct Frame {
-    pub progress: i32,
-    pub total: i32,
+    pub progress: usize,
+    pub total: usize,
     pub checked: bool,
     pub node: Node,
     pub type_t: types::Type,
 }
 
-pub fn new_frame(node: Node, type_t: types::Type, total: i32, checked: bool) -> Frame {
+pub fn new_frame(node: Node, type_t: types::Type, total: usize, checked: bool) -> Frame {
     Frame {
         progress: 0,
         total,
@@ -26,11 +26,11 @@ pub fn new_frame(node: Node, type_t: types::Type, total: i32, checked: bool) -> 
 }
 
 impl Frame {
-    pub fn get_prog(&self) -> i32 {
+    pub fn get_prog(&self) -> usize {
         self.progress
     }
 
-    pub fn set_prog(&mut self, progress: i32) {
+    pub fn set_prog(&mut self, progress: usize) {
         self.progress = progress;
     }
 
@@ -38,11 +38,11 @@ impl Frame {
         self.progress += 1;
     }
 
-    pub fn get_total(&self) -> i32 {
+    pub fn get_total(&self) -> usize {
         self.total
     }
 
-    pub fn set_total(&mut self, total: i32) {
+    pub fn set_total(&mut self, total: usize) {
         self.total = total;
     }
